requirements.txt
python-binance
python-dotenv
schedule
pandas
requests